requirements.txt
aiml==0.9.1
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-speech==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.12.1
idna==2.6
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
PyAudio==0.2.11
pyowm==2.8.0
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
spotipy==2.4.4
urllib3==1.22
